+++ app/models/user.rb
@@ -38,6 +38,7 @@ class User < ActiveRecord::Base
   has_many :exports, -> { where deleted: false }
   has_many :projects, -> { where deleted: false }
   has_many :project_favorites
+  has_many :randomization_schemes, -> { where deleted: false }
   has_many :reports, -> { where deleted: false }
   has_many :sheets, -> { where deleted: false }
   has_many :sites, -> { where deleted: false }